requirements.txt

einops==0.6.1
einops-exts==0.0.4
entrypoints==0.4
fasteners==0.18
ffmpeg==1.4
filelock==3.12.4
flamingo-pytorch==0.1.2
h5py==3.10.0
huggingface-hub==0.18.0
imageio==2.31.5
imageio-ffmpeg==0.4.8
joblib==1.2.0
numpy==1.24.4
numpy-quaternion==2022.4.3
omegaconf==2.3.0
open-clip-torch==2.20.0
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
pyparsing==3.1.1
pyrender==0.1.45
python-dateutil==2.8.2
pytorch-lightning==1.8.6
PyYAML==6.0.1
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.5
rsa==4.9
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.5.2
sentence-transformers==2.2.2
# torch==1.12.1+cu113
# torchaudio==0.12.1+cu113
# torchvision==0.13.1+cu113
tqdm==4.66.1
transformers==4.33.1
wandb==0.15.2
yarl==1.9.2
zarr==2.15.0
zipp==3.17.0
opencv-python==3.4.8.29
pytorch3d==0.3.0

hydra
moviepy
'setuptools<58.0.0'
pytest-benchmark
pytest-runner
pyhash
gym
pybullet
hydra-core
braceexpand
webdataset
matplotlib
urdfpy

